fix the namespace resource layout
